mrsapi/movie_reservation/requirements.txt
Django==5.1
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
social-auth-app-django==5.4.0
oauthlib==3.2.2
requests==2.31.0
django-redis==5.4.0
psycopg2-binary==2.9.9
gunicorn==23.0.0
django-oauth-toolkit==2.3.0